Bunch of changes, major change - Manage Orders and Update Order Status. See description for details
1. Updated Manage Orders page:
a. Only SELECT * FROM orders if no order_id was provided. Cleaned up code organization / flow
b. Used "GET" for order_id and update_status for flexibility on which page view to show depending on whether an order_id was provided or if an update was requested

2. Added 'Update Order Status' functionality
a. Created update-order-status route which performs the updating of status in the backend.
b. We use GET for this route so we can reuse the route with custom order status and order id via url_for -- i.e. we can use this when the customer wants to cancel their order in Check Order Status page
c. Status of orders presented in a drop down; user will select the new status to update to
d. If DB update successful, success message presented in Manage Orders page; failure message if otherwise

3. Added ORDER_STATUS list to the helpers.py file

4. Added JS functionality to the buttons in Manage Orders

5. Removed 'Sales' and 'Payment' (references, placeholders, etc.) to focus solely on order management / logistics

6. Updated incorrect 'read-only' attribute on input text to 'readonly'

7. Stylized TODOs on site pages